refactor: Remove plugin location wrapper abstraction layer (Phase 3)
Complete removal of pluginLocationWrappers.ts and simplify highlight extraction:

**Removed:**
- pluginLocationWrappers.ts (212 lines) - unnecessary abstraction layer
- 4 wrapper functions: findForecastLocation, findFactLocation, findSpellingErrorLocation, findHighlightLocation

**Migrated:**
- highlightExtraction/index.ts now uses findTextLocation directly
- Added local findHighlightLocation function with same logic inline
- Removed dependency on wrapper imports

**Impact:**
- Lines removed: 212
- Abstraction layers simplified: 1
- Tests: All passing ✅
- Functionality: Preserved exactly

Eliminates unnecessary middleware and improves code clarity.
